requirements.txt
docopt==0.6.2
libtmux==0.8.2
numpy==1.17.2
PyYAML==5.1.2
tmuxssh==1.1.1
